DRAW-144 refactor: phase 로직 리팩터링
