+++ core/src/main/kotlin/com/xorker/draw/mafia/phase/MafiaPhaseService.kt
@@ -16,11 +16,7 @@ internal class MafiaPhaseService(
     private val mafiaPhasePlayGameProcessor: MafiaPhasePlayGameProcessor,
     private val mafiaPhasePlayVoteProcessor: MafiaPhasePlayVoteProcessor,
     private val mafiaPhaseInferAnswerProcessor: MafiaPhaseInferAnswerProcessor,
-    private val startGameService: MafiaStartGameService,
-    private val mafiaGameService: MafiaGameService,
-    private val mafiaVoteService: MafiaVoteService,
-    private val mafiaKeywordService: MafiaKeywordService,
-    private val mafiaEndService: MafiaEndService,
+    private val mafiaPhaseEndGameProcessor: MafiaPhaseEndGameProcessor,
     private val mafiaPhaseMessenger: MafiaPhaseMessenger,
 ) : MafiaPhaseUseCase {
 
@@ -99,7 +95,7 @@ internal class MafiaPhaseService(
         val phase = synchronized(gameInfo) {
             val votePhase = gameInfo.phase
             assertIs<MafiaPhase.Vote>(votePhase)
-            mafiaEndService.endGame(gameInfo)
+            mafiaPhaseEndGameProcessor.endGame(gameInfo)
         }
 
         mafiaPhaseMessenger.broadcastPhase(gameInfo)
